Split report execution from handlers
